+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [0.6.13]
 * improvements in TiledWorldMap.
+* BREAKING CHANGES: rename mixin `WithLighting` to `Lighting`.
+* BREAKING CHANGES: rename param `tiledMap` to `map` in BonfireTiledWidget.
 * add animIdleTopLeft, animIdleBottomLeft, animIdleTopRight, animIdleBottomRight in SimplePlayer and SimpleEnemy.
 
 ## [0.6.12]